fix(camera-core): address code review issues
- FollowTarget.entityId: Types.u32 → Types.u64 to preserve EntityId generation bits
- CameraSystem: remove BigInt() cast now that entityId is already bigint (u64)
- CameraSystem: fix CameraPath first segment — use cam.x/y/z as start when prevWp is undefined
- camera-core-plugin: make setup async and await engine.use(CameraSystem)
- types: lookAt typed as Vec3 | EntityId instead of Vec3 | number
- errors: remove misleading log.error() guidance from file comment

+++ packages/camera-core/src/components.ts
@@ -46,7 +46,7 @@ export const Camera = defineComponent({
 export const FollowTarget = defineComponent({
   name: "FollowTarget",
   schema: {
-    entityId: Types.u32,
+    entityId: Types.u64,
     lerp: Types.f32,
     offsetX: Types.f32,
     offsetY: Types.f32,
